Changed URL for OWM forecast API call (#186)
* Changed URL for OWM forecast API call

Apparently, `/daily` needs to be removed for this to work with a free account, so that the query URL is just `http://api.openweathermap.org/data/2.5/forecast` and not `http://api.openweathermap.org/data/2.5/forecast/daily`. Otherwise I get a 401 "Invalid API key" response.

* Changed "cnt" to 40 as OWM returns 8 samples/day

The OWM API seems to have changed a lot since this code was originally written; they now return eight three-hourly forecast segments for each day. That's a lot of data, but there is no way to reduce it to a single daily forecast other than in client code.