Handle network errors during token refresh
Currently there is a bug when a network error occurs during obtaining access token. The error is not reported to the caller via the callback function, so the caller waits forever for the response.

Secondly this commit adds network error tolerance by retrying the failed HTTP request 4 times (resulting a total of 5 HTTP request). Furthermore it waits between retries using exponential backoff strategy according to this formula:

  delay = 100 * 9^n
  where
   - delay is the time passes between HTTP calls in milliseconds
   - n is the retry counter (starting from 0)